+++ webresearch/workflows/deep/__init__.py
@@ -1,3 +1,0 @@
-from webresearch.workflows.deep.workflow import run_deep
-
-__all__ = ["run_deep"]